SEGFAULT at cleanup, rest works

+++ fft.h
@@ -11,6 +11,6 @@ int destroy_fft( struct fft_params *fft_p, struct fft_data *fft_d );
 
 int fill_input_struct( struct fft_params *fft_p, struct fft_data *fft_d, double *in );
 
-int run_fft( struct fft_params *fft_p, struct fft_data *fft_d, double *buffer );
+int run_fft( struct fft_params *fft_p, struct fft_data *fft_d, long *buffer );
 
 #endif //FFT_H
